Use collections.Counter() to count clash roles
I didn't know off the top of my head "omg collections.Counter" would be
perfect for this.

What I _did_ know is that if I googled "python count values of a list"
that I'd find the most clean/efficient way to do it.

Turns out (and you should always suspect this at first) that python has
a built-in package that contains a method/class that helps do this.

https://docs.python.org/3/library/collections.html#collections.Counter